Updating MainMethods
This is the new MAIN commit. In this code, the program takes the players' names and letters. It also checks if the game is still going and spaces are taken or not. If nobody has won, this program also tells you if someone won or if nobody won. 

The program also cycles between players' turns and tells you which player has to take their turn. This program uses HashMaps to store the players' numbers (the players' numbers represent the spaces that each player takes on the board because all the grid spaces on the board are numbered.).

@Jayden - I couldn't find a way to modify the program so that it's able to switch the players' letters, but I find that switching the players' letters is kind of pointless, but if you know of a way to switch players' letters, then I'd appreciate it.